+++ tests/integration/test_petstore.py
@@ -18,7 +18,7 @@ from openapi_core.schema.servers.models import Server, ServerVariable
 from openapi_core.shortcuts import create_spec
 from openapi_core.validation.request.validators import RequestValidator
 from openapi_core.validation.response.validators import ResponseValidator
-from openapi_core.wrappers import MockRequest, MockResponse
+from openapi_core.wrappers.mock import MockRequest, MockResponse
 
 
 class TestPetstore(object):
