+++ tests/integration/test_wrappers.py
@@ -5,7 +5,9 @@ from werkzeug.datastructures import EnvironHeaders, ImmutableMultiDict
 from werkzeug.routing import Map, Rule, Subdomain
 from werkzeug.test import create_environ
 
-from openapi_core.wrappers import FlaskOpenAPIRequest, FlaskOpenAPIResponse
+from openapi_core.wrappers.flask import (
+    FlaskOpenAPIRequest, FlaskOpenAPIResponse,
+)
 
 
 class TestFlaskOpenAPIRequest(object):
